gtk: Move InitializingWidgetExt to subclass prelude
and rename it to CompositeTemplateInitializingExt to make it clear
it is limited to CompositeTemplate use cases
