+++ examples/list_view_apps_launcher/application_row/imp.rs
@@ -1,5 +1,4 @@
 use gtk::glib;
-use gtk::prelude::*;
 use gtk::subclass::prelude::*;
 
 use gtk::CompositeTemplate;
