+++ gtk4/Gir.toml
@@ -2576,6 +2576,9 @@ manual_traits = ["WidgetExtManual"]
         name = "region"
         const = true
     [[object.function]]
+    name = "init_template"
+    ignore = true # We manually bind this one as part of CompositeTemplatExt
+    [[object.function]]
     name = "get_name"
     rename = "widget_name"
     [[object.function]]
